Improve CSS, add link to my web page
This CSS makes the text sans-serif, and also it adds a max-width of 45rem to the body. The body is put in the horizontal centre of the html element. This should make the website look better on wide monitors (each page will have its content near the centre of the screen, instead of spread out across the whole width of a wide monitor).

Another CSS change is adding table-layout: fixed to tables. This should prevent tables overflowing their parents.

I also added a link to my promotional web page.